refactor(chat): extract ConnectorStrip, place it in the footer by the disclaimer
Move the enabled-connector icon strip out of the composer entirely into a
self-contained ConnectorStrip.vue, and render it in the footer row to the
left of the disclaimer (kept centered via balanced flex columns). Cap the
display at 5 chips (+N overflow). Composer.vue is restored to its original
state. A force refresh now bypasses any in-flight fetch and only the most
recent fetch may write the cache (generation guard).
